requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.4.1
altgraph==0.17.4
ansicon==1.89.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==24.2.0
auto-py-to-exe==2.46.0
autopy==4.0.1
awscli==1.36.12
awsebcli==3.24.1
basemap==1.4.1
basemap-data==1.3.2
bcrypt==4.3.0
blessed==1.21.0
blinker==1.9.0
boto3==1.35.70
botocore==1.35.71
bottle==0.13.3
bottle-websocket==0.2.9
branca==0.7.2
cachetools==5.5.0
cement==2.10.14
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
clr_loader==0.2.7.post0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.3
cycler==0.12.1
Cython==3.1.2
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
distlib==0.3.8
Django==5.0.7
django-cors-headers==4.5.0
django-crontab==0.7.1
django-debug-toolbar==5.2.0
django-storages==1.14.4
djangorestframework==3.15.2
docutils==0.16
Eel==0.18.1
executing==2.2.0
fabric==3.2.2
filelock==3.16.1
Flask==3.1.0
flatbuffers==24.3.25
folium==0.17.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2025.3.2
future==1.0.0
gast==0.6.0
gcloud==0.18.3
gevent==25.5.1
gevent-websocket==0.10.1
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
googleapis-common-protos==1.65.0
greenlet==3.2.2
grpcio==1.67.0
gunicorn==23.0.0
h5py==3.12.1
httplib2==0.22.0
idna==3.7
invoke==2.2.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jinxed==1.3.0
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jwcrypto==1.5.6
keras==3.6.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
mysqlclient==2.2.4
namex==0.0.8
narwhals==1.10.0
nest-asyncio==1.6.0
networkx==3.4.2
npm==0.1.1
numpy==1.26.4
oauth2client==4.1.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optional-django==0.1.0
optree==0.13.0
packaging==24.2
pandas==2.2.2
paramiko==3.5.1
parso==0.8.4
pathspec==0.12.1
pdfkit==1.0.0
pefile==2023.2.7
pillow==10.3.0
pipenv==2024.0.3
platformdirs==4.3.6
plotly==5.13.1
portalocker==3.1.1
prompt_toolkit==3.0.50
protobuf==4.25.5
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydeck==0.9.1
Pygments==2.18.0
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyJWT==2.9.0
PyKrige==1.7.2
PyNaCl==1.5.0
pyparsing==3.1.2
pypiwin32==223
pyproj==3.6.1
PyQRCode==1.2.1
Pyrebase4==4.8.0
pyshp==2.3.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-jwt==4.1.0
pythonnet==3.0.5
pytz==2024.1
pywin32==309
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.1
qrcode==8.0
razorpay==1.4.2
react==4.3.0
referencing==0.35.1
requests==2.32.3
requests-toolbelt==0.10.1
rich==13.9.3
rpds-py==0.20.0
rsa==4.7.2
s3transfer==0.10.4
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
setuptools==80.9.0
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
stack-data==0.6.3
streamlit==1.39.0
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.5.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
twilio==9.3.2
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.111
ultralytics-thop==2.0.14
urllib3==1.26.20
virtualenv==20.26.5
watchdog==5.0.3
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.9.0
wrapt==1.16.0
xyzservices==2024.6.0
yarl==1.13.1
zope.event==5.0
zope.interface==7.2
